Separated fixtures into two conftest.py files: carbon pool-specific ones in test\carbon_pools\conftest.py and general ones in test\conftest.py. Testing works fine.
