+++ test/carbon_pools/conftest.py
@@ -3,22 +3,8 @@ import numpy as np
 import os
 import pytest
 import rasterio
-import sys
-# sys.path.append('../')
-sys.path.append('/usr/local/app')
 import constants_and_names as cn
 from carbon_pools.create_carbon_pools import prepare_gain_table, mangrove_pool_ratio_dict
-
-# Deletes outputs of previous run if they exist.
-# This fixture runs only before the first parametrized run, per https://stackoverflow.com/a/62288070.
-@pytest.fixture(scope='session')
-def delete_old_outputs():
-
-    out_tests = glob.glob(f'{cn.test_data_out_dir}*.tif')
-    for f in out_tests:
-        os.remove(f)
-        print(f"Deleted {f}")
-
 
 # Makes mangrove deadwood:AGC dictionary for different continent-ecozone combinations
 @pytest.fixture(scope='session')
